Refactor

+++ app/controller/user.js
@@ -74,34 +74,6 @@ const userController = {
         
     },
 
-    /**
-     * @summary add a new user in the database
-     * @param {*} req 
-     * @param {*} res 
-     * @return {User} newUser
-     */
-    async register(req, res) {
-        // confirm password strength
-        if (req.body.password.length < 8) {
-            res.status(401).json({
-                error: 'Votre Mot de passe doit contenir 8 caracteres minimum',
-            });
-        }
-        // hash the password (https://www.npmjs.com/package/bcrypt)
-        const hashedPassword = await encrypt(req.body.password);
-        // register new user
-        const newUser = await User.create({
-            firstname: req.body.firstname,
-            lastname: req.body.lastname,
-            nickname: req.body.nickname,
-            mail: req.body.mail,
-            password: hashedPassword
-        });
-        // delete the password of the newUser before returning the newUser
-        delete newUser.password;
-        // return new user 
-        res.json(newUser);
-    },
     
     // return a list of users profils from DB 
     async getAllProfiles(req, res) {
@@ -140,6 +112,7 @@ const userController = {
                 firstname : req.body.firstname,
                 lastname: req.body.lastname,
                 nickname: req.body.nickname,
+                avatar: req.body.avatar,
                 id: id
             });
             return res.json(editedProfile);
@@ -155,11 +128,20 @@ const userController = {
             
             const foundUser = await User.getUserById(id); // Find the user via his id
             const checkedPassword = await bcrypt.compare(req.body.password, foundUser.password); // Check the current password with the one stored in the database
-            const hashedPassword = await encrypt(req.body.newPassword); // Hash the new password
-            await User.patchPwd({ // Store it in the database
-                password: hashedPassword
-            });
-            return res.json({message: "Votre mot de passe a bien été modifié"});
+            if (checkedPassword) { // If above is true and the newPassword is different than the previous one
+                if (req.body.password !== req.body.newPassword) { // If the old password is different than the new one
+                    const hashedPassword = await encrypt(req.body.newPassword); // Hash the new password
+                    await User.patchPwd({ // Store it in the database
+                        password: hashedPassword,
+                        id: id
+                    });
+                    return res.json({message: "Votre mot de passe a bien été modifié"});
+                } else {
+                    return res.status(400).json({ error: "Votre nouveau mot de passe est identique au précédent"});
+                }
+            } else {
+                return res.status(400).json({ error: "Mot de passe invalide" });
+            }
         } catch(err) {
             console.error(err);
         }
@@ -171,20 +153,23 @@ const userController = {
         try {
             // find user by id
             const userToDelete = await User.getUserById(id);
+
+            if (!userToDelete) {
+                return res.status(404).json({
+                    error: "Utilisateur inexistant"
+                })
+            }
             // test password (https://www.npmjs.com/package/bcrypt)
             const checkPassword = await bcrypt.compare(req.body.password, userToDelete.password);
 
-            console.log(checkPassword);
             // if it is not correct, return an error
             if (!checkPassword) {
                 return res.status(401).json({
                     error: "Mauvais couple email/mot de passe"
                 });
             }
-            console.log("COUCOU JE SUIS LE REQPARAMSID", id);
             // Delete the profile via his id (SQL function)
             await User.deleteProfileById(userToDelete.id);
-            console.log("COUCOU JE SUIS LE USERTODELETEID", userToDelete.id);
 
             return res.json({message: 'votre compte à bien été supprimé'})
